homiechat/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.3
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
pip-install==1.3.5
msgpack==1.0.8
pillow==10.2.0
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.1.0
python-decouple==3.8
redis==5.0.4
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
zope.interface==6.2
twilio==9.4.3